# APIdoc.yaml
openapi: 3.0.3
info:
  title: DivingRegistrationAPI
  description: API to DivingRegistration application
  version: 1.0.0
servers:
  - url: 'https://dev-restandroid.users.info.unicaen.fr/api'
  - url: 'http://127.0.0.1:8000/api'
paths:
...
  /plongees:
    get:
      description: get all active dives
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
    post:
      description: insert a new active dive
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlongeeNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
        '422':
          description: site, boat, moment, aptitude level, pilot, surface security or director don't exist, date not formatted as Y-m_d, less than 2 divers allowed, max divers less than min, pilot, security or director are not allowed
    put:
      description: update a dive
      requestBody:
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/PlongeeNoId'
      responses:
        '200':
          description: OK
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Plongee'
        '404':
          description: not found
        '422':
          description: site, boat, moment, aptitude level, pilot, surface security or director don't exist, date not formatted as Y-m_d, less than 2 divers allowed, max divers less than min, pilot, security or director are not allowed
...
components:
...
  schemas:
    Actif:
      type: object
      properties:
        actif:
          type: boolean
          description: true iff it can be chosen for new dives
          example: true
...
    Id:
      type: object
      properties:
        id:
          type: integer
          description: its unique ID
          example: 3
...
    PlongeeBase:
      type: object
      description: A dive
      properties:
        lieu:
          type: integer
          description: the ID of the site to dive at
          example: 3
        bateau:
          type: integer
          description: the ID of the boad to transport divers
          example: 1
        date:
          type: string
          format: date
          description: the date the dive will happen
          example: '2024-10-20'
        moment:
          type: integer
          description: the ID of the moment the dive will happen
          example: 1
        min_plongeurs:
          type: integer
          description: the minimum number of divers needed
          example: 4
          minimum: 2
        max_plongeurs:
          type: integer
          description: the maximum number of divers allowed. Must me greater than minimum.
          example: 4
        niveau:
          type: integer
          description: the ID of the minimum aptitude level needed for divers
          example: 10
        etat:
          type: integer
          description: the ID of the state of the dive (1 created, 2 parametrized, 3 validated, 4 done, 5 cancelled)
        securite_de_surface:
          type: integer
          description: the ID of the surface security (person)
          example: 3
        pilote:
          type: integer
          description: the ID of the pilot (person)
          example: 4
        directeur_de_plongee:
          type: integer
          description: the ID of the dive director (person)
          example: 27
    Plongee:
      allOf:
        - $ref: '#/components/schemas/Id'
        - $ref: '#/components/schemas/PlongeeBase'
        - $ref: '#/components/schemas/Actif'
    PlongeeNoId:
      allOf:
        - $ref: '#/components/schemas/PlongeeBase'
...